gcp/requirements.txt
tensorflow==2.13.1
google-cloud-storage==1.42.0
Pillow==10.2.0
flask<3.0
flask-cors==3.0.10